pyproject.toml
[project]
name = "groupchat-podcast"
version = "0.1.0"
description = "Convert iMessage group chats to podcasts using ElevenLabs TTS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=1.0.0",
    "pydub>=0.25.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
groupchat-podcast = "groupchat_podcast.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
